Refactor: move files into Models and Users folders (no logic changes)
