+++ BankApp2/Models/Checkingaccount.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using BankApp2.Users;
 
-namespace BankApp2
+namespace BankApp2.Models
 {
     public class CheckingAccount : Account
     {
